fix narrowing conversion

+++ include/engine/internal/hash_table.h
@@ -14,7 +14,7 @@ typedef struct {
     Uint32 b;
 } HT_HashTable;
 
-void *Amphora_HTGetValuePtr(const char *key, const HT_HashTable *t, size_t len);
-Uint32 Amphora_HTSetValuePtr(const char *key, Uint64 val, size_t nbytes, HT_HashTable *t, size_t len);
+void *Amphora_HTGetValuePtr(const char *key, const HT_HashTable *t, int len);
+Uint32 Amphora_HTSetValuePtr(const char *key, Uint64 val, size_t nbytes, HT_HashTable *t, int len);
 
 #endif //AMPHORA_HASH_TABLE_INTERNAL_H
